+++ application/src/main/kotlin/dev/orgsusu/application/domain/user/service/UserServiceImpl.kt
@@ -1,6 +1,6 @@
 package dev.orgsusu.application.domain.user.service
 
-import dev.orgsusu.application.domain.user.dto.UserDomainMapper
+import dev.orgsusu.application.domain.user.mapper.UserDomainMapper
 import dev.orgsusu.common.exception.CustomException
 import dev.orgsusu.application.domain.user.exception.UserExceptionDetails
 import dev.orgsusu.domain.user.model.PartialUserDomain
